% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{Directory-class}
\alias{Directory-class}
\title{Directory type}
\description{
Represents a directory input or output with a path.
}
\section{Slots}{

\describe{
\item{\code{path}}{Character string giving the directory path.}
}}

