% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Directory}
\alias{Directory}
\title{Create a Directory object}
\usage{
Directory(path)
}
\arguments{
\item{path}{Character string giving the directory path.}
}
\value{
A Directory object.
}
\description{
Create a Directory object
}
