% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{File-class}
\alias{File-class}
\title{File type}
\description{
Represents a single file input or output with a path.
}
\section{Slots}{

\describe{
\item{\code{path}}{Character string giving the file path.}
}
}
