% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{FileCollection-class}
\alias{FileCollection-class}
\title{FileCollection type}
\description{
Represents a collection of files with a character vector of paths.
}
\section{Slots}{

\describe{
\item{\code{paths}}{Character vector of file paths.}
}}

