% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{FileCollection}
\alias{FileCollection}
\title{Create a FileCollection object}
\usage{
FileCollection(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A FileCollection object.
}
\description{
Create a FileCollection object
}
