% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{infer_output_name}
\alias{infer_output_name}
\title{Infer output name from a return value's class}
\usage{
infer_output_name(value)
}
\arguments{
\item{value}{An S4 object (File, Directory, or FileCollection subclass).}
}
\value{
A character string naming the output directory.
}
\description{
Infer output name from a return value's class
}
\keyword{internal}
