% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{JsonFile}
\alias{JsonFile}
\title{Create a JsonFile object}
\usage{
JsonFile(path)
}
\arguments{
\item{path}{Character string giving the file path.}
}
\value{
A JsonFile object.
}
\description{
Create a JsonFile object
}
