% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{RasterFile}
\alias{RasterFile}
\title{Create a RasterFile object}
\usage{
RasterFile(path)
}
\arguments{
\item{path}{Character string giving the file path.}
}
\value{
A RasterFile object.
}
\description{
Create a RasterFile object
}
