% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{RasterFileCollection}
\alias{RasterFileCollection}
\title{Create a RasterFileCollection object}
\usage{
RasterFileCollection(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A RasterFileCollection object.
}
\description{
Create a RasterFileCollection object
}
