% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spade-package}
\alias{spade}
\alias{spade-package}
\title{spade: Spade Block Authoring Library for R}
\description{
Runtime library for authoring Spade blocks in R. Provides S4 types for file and directory inputs/outputs, a run() function that handles parameter loading, input scanning, handler invocation, and output writing, and a build() function for generating block manifests from function signatures.
}
\author{
\strong{Maintainer}: Spade Developers \email{dev@spade.dev}

}
\keyword{internal}
