% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade_types.R
\name{spade_types}
\alias{spade_types}
\title{Get type annotations from a function}
\usage{
spade_types(fn)
}
\arguments{
\item{fn}{A function with spade type annotations.}
}
\value{
A named list of type annotations, or NULL if none are set.
}
\description{
Retrieves the spade type annotations attached to a function via attributes.
}
