% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TabularFile}
\alias{TabularFile}
\title{Create a TabularFile object}
\usage{
TabularFile(path)
}
\arguments{
\item{path}{Character string giving the file path.}
}
\value{
A TabularFile object.
}
\description{
Create a TabularFile object
}
