% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TabularFileCollection}
\alias{TabularFileCollection}
\title{Create a TabularFileCollection object}
\usage{
TabularFileCollection(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A TabularFileCollection object.
}
\description{
Create a TabularFileCollection object
}
