% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{VectorFile}
\alias{VectorFile}
\title{Create a VectorFile object}
\usage{
VectorFile(path)
}
\arguments{
\item{path}{Character string giving the file path.}
}
\value{
A VectorFile object.
}
\description{
Create a VectorFile object
}
