% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{VectorFileCollection}
\alias{VectorFileCollection}
\title{Create a VectorFileCollection object}
\usage{
VectorFileCollection(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A VectorFileCollection object.
}
\description{
Create a VectorFileCollection object
}
