% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{write_single_output}
\alias{write_single_output}
\title{Write a single output value to outputs/\if{html}{\out{<name>}}/}
\usage{
write_single_output(name, value)
}
\arguments{
\item{name}{The output directory name.}

\item{value}{An S4 object (File, Directory, or FileCollection subclass).}
}
\description{
Write a single output value to outputs/\if{html}{\out{<name>}}/
}
\keyword{internal}
